+++ docs/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2021-05-12
+
+### Added
+
+- Ability to toggle individual features on/off
+- Amazon search links in `Buy` section of all book panes (Amazon storefront is configurable)
+
+## Fixed
+
+- Occasional `undefined` author in search links
+
 ## [1.0.1] - 2021-05-03
 
 ### Added